push code 30/03/2025

+++ CinemaMax_Server/src/main/java/org/example/cinemamax_server/repository/SubscriptionsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SubscriptionsRepository extends JpaRepository<Subscriptions, Integer> {
+public interface SubscriptionsRepository extends JpaRepository<Subscriptions, Long> {
     Optional<Subscriptions> findByName(String name);
 }
